chore: rename lib/common/stdmap.h => lib/common/StdMap.h
