+++ src/lib/base/EventQueue.h
@@ -13,7 +13,7 @@
 #include "base/IEventQueue.h"
 #include "base/PriorityQueue.h"
 #include "base/Stopwatch.h"
-#include "common/stdmap.h"
+#include "common/StdMap.h"
 #include "common/stdset.h"
 #include "mt/CondVar.h"
