+++ src/lib/platform/XWindowsKeyState.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdmap.h"
+#include "common/StdMap.h"
 #include "common/stdvector.h"
 #include "deskflow/KeyState.h"
