+++ src/lib/server/InputFilter.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/stdmap.h"
+#include "common/StdMap.h"
 #include "common/stdset.h"
 #include "deskflow/IPlatformScreen.h"
 #include "deskflow/KeyTypes.h"
